+++ CHANGELOG.md
@@ -16,7 +16,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Convert array type values to pipe (|) separated string  before make custom field value persistence. (#5933)
 - Move custom field validation to `give_checkout_error_checks` hook. (#5933)
 - Set WordPress default file max_size and allowed file type as default value of file custom field in FieldAPI. (#5933)
-
+- Custom fields have unique ID attribute in legacy consumer. (#5938
 
 ### New
 
@@ -30,6 +30,11 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Add placeholder in textarea field template in legacy consumer. (#5936)
 - Show error notice if required custom field is empty in legacy consumer. (#5936)
 
+## 2.13.2 - 2021-08-26
+
+## Fixed
+
+- Admin able to create donation form and edit existing donation form if stripe disabled. (#5935)
 
 ## 2.13.1 - 2021-08-20
 
@@ -85,6 +90,10 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Add `HTML` constant to `Give\Framework\FieldsAPI\Types` for `Html` node type (#5898)
 - Support multi-selects in Legacy Consumer select template (#5905)
+
+### Changed
+
+-   Do not extend Fields API `Form` from `Group` (#5889)
 
 ## 2.12.1 - 2021-07-22